requirements.txt
pandas
openpyxl
streamlit
pyyaml
sqlalchemy
watchdog
pytesseract
pillow
matplotlib
fastapi
uvicorn
python-multipart